Used prompt to show an input dialog

+++ resources/assets/js/projects/title.vue
@@ -24,6 +24,7 @@ export default{
             description: null,
             userId: null,
             redirectUrl: null,
+            volumes: [],
         };
     },
     computed: {
@@ -61,23 +62,39 @@ export default{
             setTimeout(() => location.href = this.redirectUrl, 2000);
         },
         deleteProject() {
-            let confirmed = confirm(`Do you really want to delete the project ${this.project.name}?`);
+            ProjectsApi.queryVolumes({ id: this.project.id }).then(
+                volumes => {
+                    this.volumes = volumes.data;
+                }
+            );
+            let inputproject = prompt(`Do you realy want to delete ${this.project.name} ?`);
+            if (inputproject.replace(/\s/g, '') == this.project.name) {
+                let confirmed = confirm(`Do you really want to delete the project ${this.project.name}?`);
 
-            if (confirmed) {
-                this.startLoading();
-                ProjectsApi.delete({id: this.project.id})
-                    .then(this.projectDeleted, this.maybeForceDeleteProject)
-                    .finally(this.finishLoading);
+                if (confirmed) {
+                    this.startLoading();
+                    ProjectsApi.delete({ id: this.project.id })
+                        .then(this.projectDeleted, this.maybeForceDeleteProject)
+                        .finally(this.finishLoading);
+                }
+
             }
         },
         maybeForceDeleteProject(response) {
             if (response.status === 400) {
-                let confirmed = confirm('Deleting this project will delete one or more volumes with all annotations! Do you want to continue?');
-                if (confirmed) {
-                    this.startLoading();
-                    ProjectsApi.delete({id: this.project.id}, {force: true})
-                        .then(this.projectDeleted, handleErrorResponse)
-                        .finally(this.finishLoading);
+                let volNames = [];
+                for (let i = 0; i < this.volumes.length; i++) {
+                    volNames[i] = this.volumes[i].name;
+                }
+                let inputVolume = prompt(`Do you realy want to delete ${volNames.toString()}?`);
+                if (inputVolume.replace(/\s/g, '') == volNames.toString().replace(/\s/g, '')) {
+                    let confirmed = confirm('Deleting this project will delete one or more volumes with all annotations! Do you want to continue?');
+                    if (confirmed) {
+                        this.startLoading();
+                        ProjectsApi.delete({ id: this.project.id }, { force: true })
+                            .then(this.projectDeleted, handleErrorResponse)
+                            .finally(this.finishLoading);
+                    }
                 }
             } else {
                 handleErrorResponse(response);
